shared/pyproject.toml
[project]
name = "shared"
version = "0.1.0"
description = "Shared benchmarks, contracts, and utilities for agents-on-azure"
requires-python = ">=3.11"

dependencies = [
    "structlog>=24.0.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "azure-monitor-opentelemetry-exporter>=1.0.0b21",
    "azure-identity>=1.15.0",
    "azure-servicebus>=7.12.0",
    "azure-eventhub>=5.11.0",
    "azure-eventgrid>=4.15.0",
    "azure-storage-queue>=12.9.0",
    "openai>=1.30.0",
    "aiohttp>=3.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["benchmarks", "contracts", "azure_clients", "utils"]
